Reorgnize nnet package
