+++ src/mygrad/nnet/activations/__init__.py
@@ -1,0 +1,7 @@
+from mygrad import tanh
+
+from .relu import relu
+from .sigmoid import sigmoid
+from .softmax import logsoftmax, softmax
+
+__all__ = ["logsoftmax", "relu", "sigmoid", "tanh", "softmax"]
