+++ src/mygrad/nnet/losses/__init__.py
@@ -1,0 +1,5 @@
+from .multiclass_hinge import multiclass_hinge
+from .softmax_crossentropy import softmax_crossentropy
+from .margin_ranking_loss import margin_ranking_loss
+
+__all__ = ["multiclass_hinge", "softmax_crossentropy", "margin_ranking_loss"]
