+++ tests/nnet/activations/test_softmax.py
