+++ tests/nnet/layers/test_batchnorm.py
