+++ tests/nnet/losses/test_hinge.py
