+++ tests/nnet/losses/test_margin_ranking.py
