+++ tests/nnet/losses/test_softmaxcrossentropy.py
